Major refactor, more sane file structure
