+++ src/runconf_ui/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "3.1.2"
+__version__ = "3.2.0"
