+++ src/runconf_ui/configuration_manager_interfaces/local_daq_conf_manager.py
@@ -1,7 +1,7 @@
 from pathlib import Path
-from runconf_ui.utils.daq_conf_management_tools.management_interface import ManagementInterface
-from runconf_ui.utils.daq_conf_tools.daq_conf_path_reader import DaqConfPathReader
-from runconf_ui.interfaces.controller.application_controller import (
+from runconf_ui.configuration_manager_interfaces.management_interface import ManagementInterface
+from runconf_ui.daq_config_interfaces.daq_config_file_io.daq_conf_path_reader import DaqConfPathReader
+from runconf_ui.runconf_ui_controllers.runconf_ui_state import (
     ShifterInterfaceState,
 )
