+++ src/runconf_ui/daq_config_interfaces/actions/action_interfaces.py
@@ -1,4 +1,4 @@
-from runconf_ui.interfaces.controller.daq_conf_wrapper import DaqConfigurationWrapper
+from runconf_ui.daq_config_interfaces.daq_config_file_io.daq_config_wrapper import DaqConfigurationWrapper
 from runconf_ui.exceptions import CiderBadActionException
 
 from abc import ABC, abstractmethod
