+++ src/runconf_ui/daq_config_interfaces/daq_config_file_io/consolidate_daq_conf.py
@@ -1,6 +1,6 @@
-from runconf_ui.interfaces.controller.daq_conf_wrapper import DaqConfigurationWrapper
+from runconf_ui.daq_config_interfaces.daq_config_file_io.daq_config_wrapper import DaqConfigurationWrapper
 
-from runconf_ui.interfaces.actions.actions import (
+from runconf_ui.daq_config_interfaces.actions.actions import (
     GetDalObjectAction,
     CopyDalAction,
     GetRelatedDalsAction,
