+++ src/runconf_ui/daq_config_interfaces/daq_tree_tools/daq_conf_tree.py
@@ -1,14 +1,14 @@
 # Essentially the tree from https://github.com/DUNE-DAQ/daqconf/blob/develop/scripts/daqconf_inspector
 
-import runconf_ui.interfaces.actions.actions as ca
-from runconf_ui.utils.shifter_config_tools.shifter_config_systems.detector_extractor import DetectorExtractor
+import runconf_ui.daq_config_interfaces.actions.actions as ca
+from runconf_ui.runconf_ui_configuration.detector_config_readers.detector_extractor import DetectorExtractor
 from runconf_ui.utils.subsystem_status import SubsystemStatus
 
 from runconf_ui.exceptions import CiderBadActionException
 
 from rich.tree import Tree
 from abc import ABC, abstractmethod
-from runconf_ui.interfaces.controller.application_controller import (
+from runconf_ui.runconf_ui_controllers.runconf_ui_state import (
     ShifterInterfaceState,
 )
