+++ src/runconf_ui/daq_config_interfaces/daq_tree_tools/daq_tree_manager.py
@@ -1,7 +1,7 @@
 from textual.screen import Screen
 from runconf_ui.widgets.multicomponent_panel import MultiComponentEnableDisablePanel
-from runconf_ui.utils.daq_conf_tools.daq_conf_tree import DaqConfTree
-from runconf_ui.interfaces.controller.application_controller import ShifterInterfaceState
+from runconf_ui.daq_config_interfaces.daq_tree_tools.daq_conf_tree import DaqConfTree
+from runconf_ui.runconf_ui_controllers.runconf_ui_state import ShifterInterfaceState
 import logging
