+++ src/runconf_ui/runconf_ui_configuration/detector_config_readers/component_extractor.py
@@ -1,5 +1,5 @@
-from runconf_ui.utils.shifter_config_tools.shifter_config_systems.extractor_interfaces import SubsystemExtractor
-import runconf_ui.interfaces.actions.actions as ca
+from runconf_ui.runconf_ui_configuration.detector_config_readers.extractor_interfaces import SubsystemExtractor
+import runconf_ui.daq_config_interfaces.actions.actions as ca
 from runconf_ui.exceptions import CiderBadActionException
 from runconf_ui.utils.subsystem_status import SubsystemStatus
