+++ src/runconf_ui/runconf_ui_configuration/detector_config_readers/detector_extractor.py
@@ -1,6 +1,6 @@
-from runconf_ui.interfaces.controller.daq_conf_wrapper import DaqConfigurationWrapper
-from runconf_ui.utils.shifter_config_tools.shifter_config_systems.extractor_interfaces import MultiItemExtractor
-from runconf_ui.utils.shifter_config_tools.shifter_config_systems.system_extractor import SystemExtractor
+from runconf_ui.daq_config_interfaces.daq_config_file_io.daq_config_wrapper import DaqConfigurationWrapper
+from runconf_ui.runconf_ui_configuration.detector_config_readers.extractor_interfaces import MultiItemExtractor
+from runconf_ui.runconf_ui_configuration.detector_config_readers.system_extractor import SystemExtractor
 from runconf_ui.exceptions import CiderBadActionException
 from runconf_ui.utils.subsystem_status import SubsystemStatus
