+++ src/runconf_ui/runconf_ui_configuration/detector_config_readers/system_extractor.py
@@ -1,8 +1,8 @@
-from runconf_ui.interfaces.controller.daq_conf_wrapper import DaqConfigurationWrapper
-from runconf_ui.utils.shifter_config_tools.shifter_config_systems.extractor_interfaces import (MultiItemExtractor,
+from runconf_ui.daq_config_interfaces.daq_config_file_io.daq_config_wrapper import DaqConfigurationWrapper
+from runconf_ui.runconf_ui_configuration.detector_config_readers.extractor_interfaces import (MultiItemExtractor,
                                                                                            SubsystemExtractor)
-from runconf_ui.utils.shifter_config_tools.shifter_config_systems.attribute_extractor import AttributeExtractor
-from runconf_ui.utils.shifter_config_tools.shifter_config_systems.component_extractor import ComponentExtractor
+from runconf_ui.runconf_ui_configuration.detector_config_readers.attribute_extractor import AttributeExtractor
+from runconf_ui.runconf_ui_configuration.detector_config_readers.component_extractor import ComponentExtractor
 from runconf_ui.utils.subsystem_status import SubsystemStatus
 from runconf_ui.exceptions import CiderBadActionException
