+++ src/runconf_ui/screens/quit_screen.py
@@ -4,10 +4,10 @@ from textual.screen import Screen
 import logging
 import os
 from pathlib import Path
-from runconf_ui.interfaces.controller.application_controller import (
+from runconf_ui.runconf_ui_controllers.runconf_ui_state import (
     ShifterInterfaceState,
 )
-from runconf_ui.utils.file_io_tools.save_file_handler import SaveFileHandler
+from runconf_ui.file_io.save_file_handler import SaveFileHandler
 
 
 class QuitScreen(Screen):
