+++ src/runconf_ui/shifter_config_files/detector_configs/np02_configuration.yml
