+++ src/runconf_ui/widgets/options_panel.py
@@ -1,13 +1,11 @@
 from runconf_ui.screens.quit_screen import QuitScreen
 from runconf_ui.screens.help_screen import HelpScreen
-from runconf_ui.widgets.popup_message import PopupMessage
 
-from textual.css.query import NoMatches
 from textual.containers import ScrollableContainer
 from textual.visual import SupportsVisual
 from textual.widgets import Button, Static
 import logging
-from runconf_ui.interfaces.controller.application_controller import (
+from runconf_ui.runconf_ui_controllers.runconf_ui_state import (
     ShifterInterfaceState,
 )
 from runconf_ui.screens.popup_manager import PopupManager
